+++ SuperMerCasa/src/main/java/com/example/supermercasa/Repositorios/RepositorioUsuario.java
@@ -5,7 +5,7 @@ import com.example.supermercasa.Entidades.Usuario;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface RepositorioUsuario extends CrudRepository<Usuario, Integer> {
+public interface RepositorioUsuario extends CrudRepository<Usuario, Long> {
 
 
 }
